apps/blender/resources/images/scripts_verifier/requirements.txt
--extra-index-url https://builds.golem.network
OpenEXR
scipy
six==1.5
Pillow==5.1.0
sklearn
numpy==1.15.4
PyWavelets
opencv-python
